refactor: ESLint 설정 분리
